# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/github/exchange": {
            "get": {
                "description": "Send a request to this endpoint to exchange the Google code you got from Google for an access token.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google Oauth exchange",
                "parameters": [
                    {
                        "type": "string",
                        "format": "code",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "state",
                        "description": "state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access token response",
                        "schema": {
                            "$ref": "#/definitions/transport.AccessTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "description": "Refresh the access token using the refresh token stored in the cookie.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "responses": {
                    "200": {
                        "description": "Access token response",
                        "schema": {
                            "$ref": "#/definitions/transport.AccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Missing cookie error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/listings/bid": {
            "post": {
                "description": "This endpoint allows users to create bids on product listings. The request body must contain the bid details in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bidding"
                ],
                "summary": "Create a bid",
                "parameters": [
                    {
                        "description": "Bid details",
                        "name": "bid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Bid"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bid created successfully with the bid ID",
                        "schema": {
                            "$ref": "#/definitions/models.CreateBidResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input or missing fields",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - failed to decode JSON payload",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listings/bids": {
            "get": {
                "description": "Retrieves a paginated list of bids for a specific product listing. The product ID, limit, and page number must be provided as query parameters.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bidding"
                ],
                "summary": "Get bids for a product listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The unique identifier of the product listing",
                        "name": "productid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The maximum number of bids to retrieve per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The page number to retrieve",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of bids for the specified product listing",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid or missing query parameters",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - product ID not provided",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listings/highest-bid": {
            "get": {
                "description": "This endpoint retrieves the highest bid for a specific product listing. The product ID must be provided as a query parameter.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bidding"
                ],
                "summary": "Get highest bid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "productid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Highest bid details",
                        "schema": {
                            "$ref": "#/definitions/models.Bid"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid or missing product ID",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - product ID not provided",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Bid": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "models.CreateBidResponse": {
            "type": "object",
            "properties": {
                "bid_id": {
                    "type": "string"
                }
            }
        },
...
        "transport.AccessTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
    }
}